+++ src/TornadoChartSettingsModel.ts
@@ -10,6 +10,9 @@ import Card = formattingSettings.SimpleCard;
 import CompositeCard = formattingSettings.CompositeCard;
 import Model = formattingSettings.Model;
 
+// Power BI represents an automatic numeric value as null, while NumUpDown.value is typed as number.
+const AutoNumericValue = null as unknown as number;
+
 import IEnumMember = powerbi.IEnumMember;
 import ILocalizationManager = powerbi.extensibility.ILocalizationManager;
 import { LegendData } from "powerbi-visuals-utils-chartutils/lib/legend/legendInterfaces";
@@ -36,10 +39,10 @@ class DataColorCardSettings extends Card {
     });
 
     name: string = TornadoObjectNames.DataPoint;
-    displayName: string = "Data colors";
-    displayNameKey: string = "Visual_DataColors";
-    description: string = "Display data color options";
-    descriptionKey: string = "Visual_Description_DataColors";
+    displayName: string = "Bars";
+    displayNameKey: string = "Visual_Bars";
+    description: string = "Display bar color options";
+    descriptionKey: string = "Visual_Description_Bars";
     slices = [this.fill];
 }
 
@@ -86,7 +89,33 @@ class CategoryAxisCardSettings extends CompositeCard {
     groups: formattingSettings.Group[] = [this.optionsGroup];
 }
 
-class NegativeBarsCardSettings extends Card {
+class NegativeBarsColorGroup extends Card {
+    constructor(slices: formattingSettings.Slice[]) {
+        super();
+        this.slices = slices;
+    }
+
+    name: string = "negativeBarsColor";
+    displayName: string = "Color";
+    displayNameKey: string = "Visual_Color";
+    slices: formattingSettings.Slice[];
+}
+
+class NegativeBarsBorderGroup extends Card {
+    constructor(slices: formattingSettings.Slice[], topLevelSlice: formattingSettings.ToggleSwitch) {
+        super();
+        this.slices = slices;
+        this.topLevelSlice = topLevelSlice;
+    }
+
+    name: string = "negativeBarsBorder";
+    displayName: string = "Border";
+    displayNameKey: string = "Visual_Border";
+    slices: formattingSettings.Slice[];
+    topLevelSlice: formattingSettings.ToggleSwitch;
+}
+
+class NegativeBarsCardSettings extends CompositeCard {
     show = new formattingSettings.ToggleSwitch({
         name: "show",
         displayName: "Show",
@@ -122,17 +151,24 @@ class NegativeBarsCardSettings extends Card {
         }
     });
 
+    showBorder = new formattingSettings.ToggleSwitch({
+        name: "showBorder",
+        displayName: "Border",
+        displayNameKey: "Visual_Border",
+        value: true
+    });
+
     borderColor = new formattingSettings.ColorPicker({
         name: "borderColor",
-        displayName: "Border color",
-        displayNameKey: "Visual_BorderColor",
+        displayName: "Color",
+        displayNameKey: "Visual_Color",
         value: { value: "" }
     });
 
     borderWidth = new formattingSettings.Slider({
         name: "borderWidth",
-        displayName: "Border width",
-        displayNameKey: "Visual_BorderWidth",
+        displayName: "Width",
+        displayNameKey: "Visual_Width",
         value: 2,
         options: {
             unitSymbol: "px",
@@ -170,22 +206,57 @@ class NegativeBarsCardSettings extends Card {
     name: string = "negativeBars";
     displayName: string = "Negative bars";
     displayNameKey: string = "Visual_NegativeBars";
-    slices = [this.fill, this.transparency, this.borderColor, this.borderWidth, this.cornerRadius];
+    colorGroup = new NegativeBarsColorGroup([this.fill, this.transparency]);
+    borderGroup = new NegativeBarsBorderGroup([this.borderColor, this.borderWidth, this.cornerRadius], this.showBorder);
+    groups: formattingSettings.Group[] = [this.colorGroup, this.borderGroup];
 }
 
-class BarAppearanceCardSettings extends Card {
+class BarAppearanceBorderGroup extends Card {
+    constructor(slices: formattingSettings.Slice[], topLevelSlice: formattingSettings.ToggleSwitch) {
+        super();
+        this.slices = slices;
+        this.topLevelSlice = topLevelSlice;
+    }
+
+    name: string = "barAppearanceBorder";
+    displayName: string = "Border";
+    displayNameKey: string = "Visual_Border";
+    slices: formattingSettings.Slice[];
+    topLevelSlice: formattingSettings.ToggleSwitch;
+}
+
+class BarAppearanceLayoutGroup extends Card {
+    constructor(slices: formattingSettings.Slice[]) {
+        super();
+        this.slices = slices;
+    }
+
+    name: string = "barAppearanceLayout";
+    displayName: string = "Layout";
+    displayNameKey: string = "Visual_Layout";
+    slices: formattingSettings.Slice[];
+}
+
+class BarAppearanceCardSettings extends CompositeCard {
+    showBorder = new formattingSettings.ToggleSwitch({
+        name: "showBorder",
+        displayName: "Border",
+        displayNameKey: "Visual_Border",
+        value: false
+    });
+
     borderColor = new formattingSettings.ColorPicker({
         name: "borderColor",
-        displayName: "Border color",
-        displayNameKey: "Visual_BorderColor",
+        displayName: "Color",
+        displayNameKey: "Visual_Color",
         value: { value: "" }
     });
 
     borderWidth = new formattingSettings.Slider({
         name: "borderWidth",
-        displayName: "Border width",
-        displayNameKey: "Visual_BorderWidth",
-        value: 0,
+        displayName: "Width",
+        displayNameKey: "Visual_Width",
+        value: 2,
         options: {
             unitSymbol: "px",
             unitSymbolAfterInput: true,
@@ -223,7 +294,7 @@ class BarAppearanceCardSettings extends Card {
         name: "barSpacing",
         displayName: "Space between bars",
         displayNameKey: "Visual_BarSpacing",
-        value: 0,
+        value: 16,
         options: {
             unitSymbol: "%",
             unitSymbolAfterInput: true,
@@ -241,7 +312,9 @@ class BarAppearanceCardSettings extends Card {
     name: string = "barAppearance";
     displayName: string = "Bar appearance";
     displayNameKey: string = "Visual_BarAppearance";
-    slices = [this.borderColor, this.borderWidth, this.cornerRadius, this.barSpacing];
+    borderGroup = new BarAppearanceBorderGroup([this.borderColor, this.borderWidth, this.cornerRadius], this.showBorder);
+    layoutGroup = new BarAppearanceLayoutGroup([this.barSpacing]);
+    groups: formattingSettings.Group[] = [this.borderGroup, this.layoutGroup];
 }
 
 class CenterLineCardSettings extends Card {
@@ -298,14 +371,14 @@ class ChartAreaCardSettings extends Card {
 
     backgroundColor = new formattingSettings.ColorPicker({
         name: "backgroundColor",
-        displayName: "Background color",
-        displayNameKey: "Visual_BackgroundColor",
+        displayName: "Color",
+        displayNameKey: "Visual_Color",
         value: { value: "" }
     });
 
     name: string = "chartArea";
-    displayName: string = "Chart Area";
-    displayNameKey: string = "Visual_ChartArea";
+    displayName: string = "Plot area background";
+    displayNameKey: string = "Visual_PlotAreaBackground";
     slices = [this.backgroundColor];
 }
 
@@ -325,14 +398,14 @@ export enum LabelPosition {
 
 export const contentOptions: IEnumMemberWithDisplayNameKey[] = [
     { value: LabelDisplayMode.Value, displayName: "Value", key: "Visual_Value" },
-    { value: LabelDisplayMode.Percentage, displayName: "Percentage", key: "Visual_Percentage" },
+    { value: LabelDisplayMode.Percentage, displayName: "%", key: "Visual_Percentage" },
     { value: LabelDisplayMode.ValueAndPercentage, displayName: "Value (%)", key: "Visual_ValueAndPercentage" },
 ];
 
 export const dataLabelPositionOptions: IEnumMemberWithDisplayNameKey[] = [
     { value: LabelPosition.Auto, displayName: "Auto", key: "Visual_Position_Auto" },
-    { value: LabelPosition.OutsideEnd, displayName: "Outside end", key: "Visual_Position_OutsideEnd" },
     { value: LabelPosition.InsideEnd, displayName: "Inside end", key: "Visual_Position_InsideEnd" },
+    { value: LabelPosition.OutsideEnd, displayName: "Outside end", key: "Visual_Position_OutsideEnd" },
     { value: LabelPosition.InsideCenter, displayName: "Inside center", key: "Visual_Position_InsideCenter" },
     { value: LabelPosition.InsideBase, displayName: "Inside base", key: "Visual_Position_InsideBase" },
 ];
@@ -365,9 +438,9 @@ class LabelsValuesGroup extends Card {
 
     labelPrecision = new formattingSettings.NumUpDown({
         name: "labelPrecision",
-        displayName: "Decimal Places",
+        displayName: "Decimal places",
         displayNameKey: "Visual_DataLabels_DecimalPlaces",
-        value: 0,
+        value: AutoNumericValue,
         options: {
             minValue: {
                 type: powerbiVisualsApi.visuals.ValidatorType.Min,
@@ -380,9 +453,26 @@ class LabelsValuesGroup extends Card {
         }
     });
 
+    percentagePrecision = new formattingSettings.NumUpDown({
+        name: "percentagePrecision",
+        displayName: "Percentage decimal places",
+        displayNameKey: "Visual_DataLabels_PercentageDecimalPlaces",
+        value: AutoNumericValue,
+        options: {
+            minValue: {
+                type: powerbiVisualsApi.visuals.ValidatorType.Min,
+                value: 0,
+            },
+            maxValue: {
+                type: powerbiVisualsApi.visuals.ValidatorType.Max,
+                value: 10,
+            }
+        }
+    });
+
     labelDisplayUnits = new formattingSettings.AutoDropdown({
         name: "labelDisplayUnits",
-        displayName: "Display Units",
+        displayName: "Display units",
         displayNameKey: "Visual_DisplayUnits",
         value: 1
     });
@@ -411,7 +501,19 @@ class LabelsValuesGroup extends Card {
     name: string = "values";
     displayName: string = "Values";
     displayNameKey: string = "Visual_Values";
-    slices = [this.font, this.labelPrecision, this.labelDisplayUnits, this.insideFill, this.outsideFill, this.negativeFill];
+    slices = [this.font, this.labelPrecision, this.percentagePrecision, this.labelDisplayUnits];
+}
+
+class LabelsColorGroup extends Card {
+    constructor(valuesGroup: LabelsValuesGroup) {
+        super();
+        this.slices = [valuesGroup.insideFill, valuesGroup.outsideFill, valuesGroup.negativeFill];
+    }
+
+    name: string = "color";
+    displayName: string = "Color";
+    displayNameKey: string = "Visual_Color";
+    slices: formattingSettings.Slice[];
 }
 
 export class DataLabelSettings extends CompositeCard {
@@ -426,11 +528,12 @@ export class DataLabelSettings extends CompositeCard {
 
     public labelsOptionsGroup: LabelsOptionsGroup = new LabelsOptionsGroup();
     public labelsValuesGroup: LabelsValuesGroup = new LabelsValuesGroup();
+    public labelsColorGroup: LabelsColorGroup = new LabelsColorGroup(this.labelsValuesGroup);
 
     name: string = "labels";
-    displayName: string = "Data Labels";
+    displayName: string = "Data labels";
     displayNameKey: string = "Visual_DataLabels";
-    groups: formattingSettings.Group[] = [this.labelsOptionsGroup, this.labelsValuesGroup];
+    groups: formattingSettings.Group[] = [this.labelsOptionsGroup, this.labelsValuesGroup, this.labelsColorGroup];
 }
 
 interface IEnumMemberWithDisplayNameKey extends IEnumMember{
@@ -442,10 +545,10 @@ const positionOptions : IEnumMemberWithDisplayNameKey[] = [
     {value : LegendPosition[LegendPosition.Bottom], displayName : "Bottom", key: "Visual_Legend_Position_Bottom"},
     {value : LegendPosition[LegendPosition.Left], displayName : "Left", key: "Visual_Legend_Position_Left"}, 
     {value : LegendPosition[LegendPosition.Right], displayName : "Right", key: "Visual_Legend_Position_Right"}, 
-    {value : LegendPosition[LegendPosition.TopCenter], displayName : "Top Center", key: "Visual_Legend_Position_Top_Center"}, 
-    {value : LegendPosition[LegendPosition.BottomCenter], displayName : "Bottom Center", key: "Visual_Legend_Position_Bottom_Center"}, 
-    {value : LegendPosition[LegendPosition.LeftCenter], displayName : "Left Center", key: "Visual_Legend_Position_Left_Center"}, 
-    {value : LegendPosition[LegendPosition.RightCenter], displayName : "Right Center", key: "Visual_Legend_Position_Right_Center"}, 
+    {value : LegendPosition[LegendPosition.TopCenter], displayName : "Top center", key: "Visual_Legend_Position_Top_Center"},
+    {value : LegendPosition[LegendPosition.BottomCenter], displayName : "Bottom center", key: "Visual_Legend_Position_Bottom_Center"},
+    {value : LegendPosition[LegendPosition.LeftCenter], displayName : "Left center", key: "Visual_Legend_Position_Left_Center"},
+    {value : LegendPosition[LegendPosition.RightCenter], displayName : "Right center", key: "Visual_Legend_Position_Right_Center"},
 ];
 
 class BaseFontCardSettings extends formattingSettings.FontControl {
@@ -458,18 +561,19 @@ class BaseFontCardSettings extends formattingSettings.FontControl {
     public static defaultFontFamily: string = "wf_standard-font, helvetica, arial, sans-serif";
     public static minFontSize: number = 8;
     public static maxFontSize: number = 60;
-    constructor(defaultFontSize: number, settingName: string = ""){
+    constructor(defaultFontSize: number, settingName: string = "", fontFamily: string = BaseFontCardSettings.defaultFontFamily){
         super(
             new formattingSettings.FontControl({
                 name: BaseFontCardSettings.fontName + settingName,
+                displayName: "Font",
                 displayNameKey: "Visual_FontControl",
                 fontFamily: new formattingSettings.FontPicker({
                     name: BaseFontCardSettings.fontFamilyName + settingName,
-                    value: BaseFontCardSettings.defaultFontFamily
+                    value: fontFamily
                 }),
                 fontSize: new formattingSettings.NumUpDown({
                     name: BaseFontCardSettings.fontSizeName + settingName,
-                    displayNameKey: "Visual_FontSize",
+                    displayNameKey: "Visual_TextSize",
                     value: defaultFontSize,
                     options: {
                         minValue: {
@@ -517,15 +621,16 @@ class LegendOptionsGroup extends Card {
 
 class LegendTextGroup extends Card {
     public defaultLabelColor: string = "";
-    public defaultFontSize: number = 8;
+    public defaultFontSize: number = 9;
 
     public labelColor = new formattingSettings.ColorPicker({
         name: "labelColor",
-        displayNameKey: "Visual_LabelColor",
+        displayName: "Color",
+        displayNameKey: "Visual_Color",
         value: { value: this.defaultLabelColor },
     });
 
-    public font = new BaseFontCardSettings(this.defaultFontSize);
+    public font = new BaseFontCardSettings(this.defaultFontSize, "", "Segoe UI");
 
     name: string = "legendText";
     displayName: string = "Text";
@@ -547,9 +652,10 @@ class LegendTitleGroup extends Card {
 
     public titleText = new formattingSettings.TextInput({
         name: "titleText",
-        displayNameKey: "Visual_TitleText",
+        displayName: "Text",
+        displayNameKey: "Visual_Text",
         value: this.defaultTitleText,
-        placeholder: "Title text",
+        placeholder: "Text",
     });
 
     name: string = TornadoObjectNames.LegendTitle;
@@ -567,7 +673,7 @@ export class LegendCardSettings extends CompositeCard {
 
     public show = new formattingSettings.ToggleSwitch({
         name: "show",
-        displayNameKey: "Visual_LegendShow",
+        displayNameKey: "Visual_Legend_Show",
         value: this.defaultShow,
     });
 
@@ -596,13 +702,15 @@ export class BaseFontControlSettings extends formattingSettings.FontControl {
         super(
             new formattingSettings.FontControl({
                 name: "font",
+                displayName: "Font",
+                displayNameKey: "Visual_FontControl",
                 fontFamily: new formattingSettings.FontPicker({
                     name: "fontFamily",
                     value: FontDefaultOptions.DefaultFontFamily
                 }),
                 fontSize: new formattingSettings.NumUpDown({
                     name: "fontSize",
-                    displayName: "Text Size",
+                    displayName: "Text size",
                     displayNameKey: "Visual_TextSize",
                     value: defaultFontSize,
                     options: {
@@ -633,7 +741,31 @@ export class BaseFontControlSettings extends formattingSettings.FontControl {
     }
 }
 
-export class CategoryCardSettings extends Card {
+class CategoryOptionsGroup extends Card {
+    constructor(position: formattingSettings.ItemDropdown) {
+        super();
+        this.slices = [position];
+    }
+
+    name: string = "categoryOptions";
+    displayName: string = "Options";
+    displayNameKey: string = "Visual_Options";
+    slices: formattingSettings.Slice[];
+}
+
+class CategoryTextGroup extends Card {
+    constructor(font: formattingSettings.FontControl, fill: formattingSettings.ColorPicker) {
+        super();
+        this.slices = [font, fill];
+    }
+
+    name: string = "categoryText";
+    displayName: string = "Text";
+    displayNameKey: string = "Visual_Text";
+    slices: formattingSettings.Slice[];
+}
+
+export class CategoryCardSettings extends CompositeCard {
     show = new formattingSettings.ToggleSwitch({
         name: "show",
         displayName: "Show",
@@ -643,7 +775,7 @@ export class CategoryCardSettings extends Card {
     
     topLevelSlice? = this.show;
 
-    font: formattingSettings.FontControl = new BaseFontControlSettings(8);
+    font: formattingSettings.FontControl = new BaseFontControlSettings(9);
 
     fill = new formattingSettings.ColorPicker({
         name: "fill",
@@ -663,7 +795,9 @@ export class CategoryCardSettings extends Card {
     name: string = TornadoObjectNames.Categories;
     displayName: string = "Group";
     displayNameKey: string = "Visual_Group";
-    slices = [this.positionDropdown, this.font, this.fill];
+    optionsGroup = new CategoryOptionsGroup(this.positionDropdown);
+    textGroup = new CategoryTextGroup(this.font, this.fill);
+    groups: formattingSettings.Group[] = [this.optionsGroup, this.textGroup];
 }
 
 
@@ -680,10 +814,10 @@ export class TornadoChartSettingsModel extends Model {
 
     cards = [
         this.dataColors,
-        this.categoryAxis,
         this.barAppearance,
         this.negativeBars,
         this.centerLine,
+        this.categoryAxis,
         this.dataLabels,
         this.legend,
         this.category,
@@ -705,6 +839,16 @@ export class TornadoChartSettingsModel extends Model {
 
     public setVisibilityOfLegendCardSettings(legend: LegendData){
         this.legend.visible = legend.dataPoints.length > 0;
+    }
+
+    public updateDataLabelControlsState(): void {
+        const displayMode = this.dataLabels.labelsOptionsGroup.displayFormat.value.value;
+        const includesValue = displayMode !== LabelDisplayMode.Percentage;
+        const includesPercentage = displayMode !== LabelDisplayMode.Value;
+
+        this.dataLabels.labelsValuesGroup.labelPrecision.disabled = !includesValue;
+        this.dataLabels.labelsValuesGroup.labelDisplayUnits.disabled = !includesValue;
+        this.dataLabels.labelsValuesGroup.percentagePrecision.disabled = !includesPercentage;
     }
     
     public populateDataColorSlice(dataPoints: TornadoChartSeries[]){